requirements.txt
psycopg2-binary==2.9.9
pybind11
python-dotenv
pandas
pymystem3
scikit-learn
elasticsearch
pydantic
aiohttp
uvicorn
fastapi
fastapi-utils
urllib3
sentence-transformers
gensim
scikit-learn
